refactor cli project
