+++ source-generator/Cli/IProject.cs
@@ -1,0 +1,6 @@
+namespace Cli;
+
+public interface IProject
+{
+    void Do(string input, string output);
+}
