+++ source-generator/Cli/Program.cs
@@ -1,30 +1,9 @@
-﻿using System.Text.RegularExpressions;
+﻿using Cli;
 
-var process = args[0];
+var projectName = args[0];
 var input = args[1];
 var output = args[2];
 
-if(process == "domain")
-{
-    var generatedCode = File.ReadAllText(input);
+ProjectFactory projectFactory = new(projectName);
 
-    string pattern = @"===JSON BEGIN===\s*(.*?)\s*===JSON END===";
-    Match match = Regex.Match(generatedCode, pattern);
-
-    string capturedText = "";
-
-    if (match.Success)
-    {
-        capturedText = match.Groups[1].Value;
-    }
-
-    capturedText = capturedText.Replace("'", "\"");
-
-    File.WriteAllText(output, capturedText);
-}
-else if (process == "webapp")
-{
-    var jsonFile = File.ReadAllText(input);
-
-    File.WriteAllText(output, jsonFile);
-}
+projectFactory.GetClass().Do(input, output);
